greybus: use uint16_t for cport ids in handlers
Update all the protocol handlers so they use uint16_t to represent
the CPort IDs they work with.  Rename all CPort ID variables to be
"cport_id" (not "cport").

Update format specifiers to use "%hu" rather than "%d" when printing
CPort IDs.

+++ gbsim.h
@@ -132,7 +132,7 @@ static inline void gbsim_dump(__u8 *buf, size_t size)
 	fprintf(stdout, "\n");
 }
 
-static inline int cport_to_module_id(unsigned int cport)
+static inline int cport_to_module_id(uint16_t cport)
 {
 	/* FIXME can identify based on registered cport module */
 	return 1;
@@ -157,17 +157,17 @@ void send_link_up(int, int);
 void *recv_thread(void *);
 void recv_thread_cleanup(void *);
 
-void gpio_handler(unsigned int, void *, size_t);
+void gpio_handler(uint16_t, void *, size_t);
 void gpio_init(void);
 
-void i2c_handler(unsigned int, void *, size_t);
+void i2c_handler(uint16_t, void *, size_t);
 void i2c_init(void);
 
-void pwm_handler(unsigned int, void *, size_t);
+void pwm_handler(uint16_t, void *, size_t);
 void pwm_init(void);
 
-void i2s_mgmt_handler(unsigned int, void *, size_t);
-void i2s_data_handler(unsigned int, void *, size_t);
+void i2s_mgmt_handler(uint16_t, void *, size_t);
+void i2s_data_handler(uint16_t, void *, size_t);
 void i2s_init(void);
 
 bool manifest_parse(void *data, size_t size);
